05_django/10_django_rest_framework/django_ws_10_1/artists/views.py
from django.shortcuts import render
from rest_framework import status
from rest_framework.response import Response
from rest_framework.decorators import api_view

from .serializers import ArtistSerializers
# Create your views here.
@api_view(['POST'])
def create_artist(request):
    if request.method == 'POST':
        serializers = ArtistSerializers(data = request.data)
        if serializers.is_valid():
            serializers.save()
            return Response(serializers.data, status=status.HTTP_201_CREATED)
        return Response(serializers.errors, status=status.HTTP_400_BAD_REQUEST)