05_django/10_django_rest_framework/django_ws_10_1/requirements.txt
asgiref==3.7.2
Django==4.2.2
djangorestframework==3.15.1
sqlparse==0.4.4
typing_extensions==4.6.3
tzdata==2023.3
